feat: implement telemetry metrics with counters, gauges, and histograms in the API (#1041)
