+++ core/src/ten_rust/src/endpoint/mod.rs
@@ -4,8 +4,8 @@
 // Licensed under the Apache License, Version 2.0, with certain conditions.
 // Refer to the "LICENSE" file in the root directory for more information.
 //
-mod api_handler;
-mod telemetry_handler;
+mod api;
+mod telemetry;
 
 use std::os::raw::c_char;
 use std::ptr;
@@ -16,10 +16,7 @@ use anyhow::Result;
 use futures::channel::oneshot;
 use futures::future::select;
 use futures::FutureExt;
-use prometheus::{
-    Counter, CounterVec, Gauge, GaugeVec, Histogram, HistogramOpts,
-    HistogramVec, Opts, Registry,
-};
+use prometheus::Registry;
 
 use crate::constants::{
     ENDPOINT_SERVER_BIND_MAX_RETRIES, ENDPOINT_SERVER_BIND_RETRY_INTERVAL_SECS,
@@ -35,24 +32,15 @@ pub struct TelemetrySystem {
     actix_shutdown_tx: Option<oneshot::Sender<()>>,
 }
 
-pub enum MetricHandle {
-    Counter(Counter),
-    CounterVec(CounterVec),
-    Gauge(Gauge),
-    GaugeVec(GaugeVec),
-    Histogram(Histogram),
-    HistogramVec(HistogramVec),
-}
-
 /// Configure API endpoints.
 fn configure_routes(cfg: &mut web::ServiceConfig, registry: Registry) {
     let registry_clone = registry;
 
     // Configure telemetry endpoint.
-    telemetry_handler::configure_telemetry_route(cfg, registry_clone.clone());
+    telemetry::configure_telemetry_route(cfg, registry_clone.clone());
 
     // Configure API endpoints.
-    api_handler::configure_api_route(cfg);
+    api::configure_api_route(cfg);
 }
 
 /// Creates an HTTP server with retry mechanism if binding fails.
@@ -121,17 +109,100 @@ fn create_endpoint_server_with_retry(
     Some(server_builder.run())
 }
 
+/// Creates and starts a server thread that runs the actix system with the
+/// provided server.
+///
+/// This function encapsulates the logic for running an actix server in a
+/// separate thread, handling both normal operation and shutdown requests.
+fn create_server_thread(
+    server: actix_web::dev::Server,
+) -> (thread::JoinHandle<()>, oneshot::Sender<()>) {
+    // Get a handle to the server to control it later.
+    let server_handle = server.handle();
+
+    // Create a channel to send shutdown signals to the server thread.
+    let (shutdown_tx, shutdown_rx) = oneshot::channel::<()>();
+
+    // Spawn a new thread to run the actix system.
+    let server_thread_handle = thread::spawn(move || {
+        // Create a new actix system.
+        let system = actix_rt::System::new();
+
+        // Block on the async executor to run our server and shutdown logic.
+        let result: Result<()> = system.block_on(async move {
+            // Set up the concurrent execution of server and shutdown tasks.
+
+            // The server task handles normal operation and error reporting.
+            let server_future = async {
+                match server.await {
+                    Ok(_) => {
+                        // Server completed normally (unlikely).
+                        eprintln!("Endpoint server completed normally");
+                    }
+                    Err(e) => {
+                        // Server encountered an error.
+                        eprintln!("Endpoint server error: {e}");
+                        // Force the entire process to exit immediately.
+                        std::process::exit(-1);
+                    }
+                }
+            }
+            .fuse();
+
+            // The shutdown task waits for a signal to gracefully stop the
+            // server.
+            let shutdown_future = async move {
+                // Wait for shutdown signal.
+                let _ = shutdown_rx.await;
+
+                eprintln!("Shutting down endpoint server (graceful stop)...");
+
+                // Gracefully stop the server.
+                server_handle.stop(true).await;
+
+                // Terminate the actix system after the server is fully down.
+                actix_rt::System::current().stop();
+            }
+            .fuse();
+
+            // Use `futures::select!` to concurrently execute both futures
+            // and respond to whichever completes first.
+            futures::pin_mut!(server_future, shutdown_future);
+            select(server_future, shutdown_future).await;
+
+            eprintln!("Endpoint server shut down.");
+            Ok(())
+        });
+
+        // Handle any errors from the actix system.
+        if let Err(e) = result {
+            eprintln!("Fatal error in endpoint server thread: {:?}", e);
+            std::process::exit(-1);
+        }
+    });
+
+    (server_thread_handle, shutdown_tx)
+}
+
 /// Initialize the endpoint system.
+///
+/// # Safety
+///
+/// This function takes a raw C string pointer. The pointer must be valid and
+/// point to a properly null-terminated string. The returned pointer must be
+/// freed with `ten_endpoint_system_shutdown` to avoid memory leaks.
 #[no_mangle]
-#[allow(clippy::not_unsafe_ptr_arg_deref)]
-pub extern "C" fn ten_endpoint_system_create(
+pub unsafe extern "C" fn ten_endpoint_system_create(
     endpoint: *const c_char,
 ) -> *mut TelemetrySystem {
-    let endpoint_str = match unsafe { CStr::from_ptr(endpoint) }.to_str() {
+    // Safely convert C string to Rust string.
+    let endpoint_str = match CStr::from_ptr(endpoint).to_str() {
         Ok(s) if !s.trim().is_empty() => s.to_string(),
         _ => return ptr::null_mut(),
     };
 
+    // Create a new Prometheus registry.
+    //
     // Note: `prometheus::Registry` internally uses `Arc` and `RwLock` to
     // achieve thread safety, so there is no need to add additional locking
     // mechanisms. It can be used directly here.
@@ -146,623 +217,68 @@ pub extern "C" fn ten_endpoint_system_create(
         None => return ptr::null_mut(),
     };
 
-    let server_handle = server.handle();
+    // Create the server thread and get the shutdown channel.
+    let (server_thread_handle, shutdown_tx) = create_server_thread(server);
 
-    // Create an `oneshot` channel to notify the actix system to shut down.
-    let (shutdown_tx, shutdown_rx) = oneshot::channel::<()>();
-
-    let server_thread_handle = thread::spawn(move || {
-        let sys = actix_rt::System::new();
-
-        // Use `sys.block_on(...)` to execute an async block.
-        let result: Result<()> = sys.block_on(async move {
-            // Wait for either the server to finish (unlikely) or the shutdown
-            // signal.
-            let server_future = async {
-                if let Err(e) = server.await {
-                    eprintln!("Endpoint server error: {e}");
-                    // Force the entire process to exit immediately.
-                    std::process::exit(-1);
-                }
-            }
-            .fuse();
-
-            // Instead of just calling `System::current().stop()`, we call
-            // `server_handle.stop(true).await` to shut down the server
-            // gracefully, then stop the system.
-            let shutdown_future = async move {
-                let _ = shutdown_rx.await;
-
-                eprintln!("Shutting down endpoint server (graceful stop)...");
-                server_handle.stop(true).await;
-
-                // The server is actually stopped (socket closed).
-
-                // Terminates the actix system after the server is fully down.
-                actix_rt::System::current().stop();
-            }
-            .fuse();
-
-            // Use `futures::select!` to concurrently execute two futures.
-            futures::pin_mut!(server_future, shutdown_future);
-            select(server_future, shutdown_future).await;
-
-            eprintln!("Endpoint server shut down.");
-            Ok(())
-        });
-
-        if let Err(e) = result {
-            eprintln!("Fatal error in endpoint server thread: {:?}", e);
-
-            std::process::exit(-1);
-        }
-    });
-
-    let actix_thread = Some(server_thread_handle);
-
+    // Create and return the TelemetrySystem.
     let system = TelemetrySystem {
         registry,
-        actix_thread,
+        actix_thread: Some(server_thread_handle),
         actix_shutdown_tx: Some(shutdown_tx),
     };
 
+    // Convert to raw pointer for C API.
     Box::into_raw(Box::new(system))
 }
 
 /// Shut down the endpoint system, stop the server, and clean up all resources.
+///
+/// # Safety
+///
+/// This function assumes that `system_ptr` is either null or a valid pointer to
+/// a `TelemetrySystem` that was previously created with
+/// `ten_endpoint_system_create`. Calling this function with an invalid pointer
+/// will lead to undefined behavior.
 #[no_mangle]
-#[allow(clippy::not_unsafe_ptr_arg_deref)]
-pub extern "C" fn ten_endpoint_system_shutdown(
+pub unsafe extern "C" fn ten_endpoint_system_shutdown(
     system_ptr: *mut TelemetrySystem,
 ) {
     debug_assert!(!system_ptr.is_null(), "System pointer is null");
-
+    // Early return for null pointers.
     if system_ptr.is_null() {
+        eprintln!("Warning: Attempt to shut down null TelemetrySystem pointer");
         return;
     }
 
-    // Retrieve ownership using `Box::from_raw`, and it will be automatically
-    // released when the function exits.
-    let system = unsafe { Box::from_raw(system_ptr) };
+    // Retrieve ownership using `Box::from_raw`. This transfers ownership to
+    // Rust, and the Box will be automatically dropped when it goes out of
+    // scope.
+    let system = Box::from_raw(system_ptr);
 
     // Notify the actix system to shut down through the `oneshot` channel.
     if let Some(shutdown_tx) = system.actix_shutdown_tx {
         eprintln!("Shutting down endpoint server...");
-        let _ = shutdown_tx.send(());
+        if let Err(e) = shutdown_tx.send(()) {
+            eprintln!("Failed to send shutdown signal: {:?}", e);
+
+            panic!("Failed to send shutdown signal");
+        }
+    } else {
+        eprintln!("No shutdown channel available for the endpoint server");
     }
 
+    // Wait for the server thread to complete.
     if let Some(server_thread_handle) = system.actix_thread {
         eprintln!("Waiting for endpoint server to shut down...");
-        let _ = server_thread_handle.join();
-    }
-}
-
-fn convert_label_names(
-    names_ptr: *const *const c_char,
-    names_len: usize,
-) -> Option<Vec<String>> {
-    if names_ptr.is_null() {
-        return Some(vec![]);
-    }
-
-    let mut result = Vec::with_capacity(names_len);
-
-    for i in 0..names_len {
-        let c_str_ptr = unsafe { *names_ptr.add(i) };
-        if c_str_ptr.is_null() {
-            return None;
-        }
-
-        let c_str = unsafe { CStr::from_ptr(c_str_ptr) };
-        match c_str.to_str() {
-            Ok(s) => result.push(s.to_string()),
-            Err(_) => return None,
-        }
-    }
-
-    Some(result)
-}
-
-fn convert_label_values(
-    values_ptr: *const *const c_char,
-    values_len: usize,
-) -> Option<Vec<String>> {
-    if values_ptr.is_null() {
-        return Some(vec![]);
-    }
-
-    let mut result = Vec::with_capacity(values_len);
-
-    for i in 0..values_len {
-        let c_str_ptr = unsafe { *values_ptr.add(i) };
-        if c_str_ptr.is_null() {
-            return None;
-        }
-
-        let c_str = unsafe { CStr::from_ptr(c_str_ptr) };
-        match c_str.to_str() {
-            Ok(s) => result.push(s.to_string()),
-            Err(_) => return None,
-        }
-    }
-
-    Some(result)
-}
-
-fn create_metric_counter(
-    system: &mut TelemetrySystem,
-    name_str: &str,
-    help_str: &str,
-) -> Result<MetricHandle> {
-    let counter_opts = Opts::new(name_str, help_str);
-    match Counter::with_opts(counter_opts) {
-        Ok(counter) => {
-            if let Err(e) = system.registry.register(Box::new(counter.clone()))
-            {
-                eprintln!("Error registering counter: {:?}", e);
-                return Err(anyhow::anyhow!("Error registering counter"));
-            }
-            Ok(MetricHandle::Counter(counter))
-        }
-        Err(_) => Err(anyhow::anyhow!("Error creating counter")),
-    }
-}
-
-fn create_metric_counter_with_labels(
-    system: &mut TelemetrySystem,
-    name_str: &str,
-    help_str: &str,
-    label_names: &[&str],
-) -> Result<MetricHandle> {
-    let counter_opts = Opts::new(name_str, help_str);
-    match CounterVec::new(counter_opts, label_names) {
-        Ok(counter_vec) => {
-            if let Err(e) =
-                system.registry.register(Box::new(counter_vec.clone()))
-            {
-                eprintln!("Error registering counter vec: {:?}", e);
-                return Err(anyhow::anyhow!("Error registering counter"));
-            }
-            Ok(MetricHandle::CounterVec(counter_vec))
-        }
-        Err(_) => Err(anyhow::anyhow!("Error creating counter")),
-    }
-}
-
-fn create_metric_gauge(
-    system: &mut TelemetrySystem,
-    name_str: &str,
-    help_str: &str,
-) -> Result<MetricHandle> {
-    let gauge_opts = Opts::new(name_str, help_str);
-    match Gauge::with_opts(gauge_opts) {
-        Ok(gauge) => {
-            if let Err(e) = system.registry.register(Box::new(gauge.clone())) {
-                eprintln!("Error registering gauge: {:?}", e);
-                return Err(anyhow::anyhow!("Error registering gauge"));
-            }
-            Ok(MetricHandle::Gauge(gauge))
-        }
-        Err(_) => Err(anyhow::anyhow!("Error creating gauge")),
-    }
-}
-
-fn create_metric_gauge_with_labels(
-    system: &mut TelemetrySystem,
-    name_str: &str,
-    help_str: &str,
-    label_names: &[&str],
-) -> Result<MetricHandle> {
-    let gauge_opts = Opts::new(name_str, help_str);
-    match GaugeVec::new(gauge_opts, label_names) {
-        Ok(gauge_vec) => {
-            if let Err(e) =
-                system.registry.register(Box::new(gauge_vec.clone()))
-            {
-                eprintln!("Error registering gauge vec: {:?}", e);
-                return Err(anyhow::anyhow!("Error registering gauge"));
-            }
-            Ok(MetricHandle::GaugeVec(gauge_vec))
-        }
-        Err(_) => Err(anyhow::anyhow!("Error creating gauge")),
-    }
-}
-
-fn create_metric_histogram(
-    system: &mut TelemetrySystem,
-    name_str: &str,
-    help_str: &str,
-) -> Result<MetricHandle> {
-    let hist_opts = HistogramOpts::new(name_str, help_str);
-    match Histogram::with_opts(hist_opts) {
-        Ok(histogram) => {
-            if let Err(e) =
-                system.registry.register(Box::new(histogram.clone()))
-            {
-                eprintln!("Error registering histogram: {:?}", e);
-                return Err(anyhow::anyhow!("Error registering histogram"));
-            }
-            Ok(MetricHandle::Histogram(histogram))
-        }
-        Err(_) => Err(anyhow::anyhow!("Error creating histogram")),
-    }
-}
-
-fn create_metric_histogram_with_labels(
-    system: &mut TelemetrySystem,
-    name_str: &str,
-    help_str: &str,
-    label_names: &[&str],
-) -> Result<MetricHandle> {
-    let hist_opts = HistogramOpts::new(name_str, help_str);
-    match HistogramVec::new(hist_opts, label_names) {
-        Ok(histogram_vec) => {
-            if let Err(e) =
-                system.registry.register(Box::new(histogram_vec.clone()))
-            {
-                eprintln!("Error registering histogram vec: {:?}", e);
-                return Err(anyhow::anyhow!("Error registering histogram"));
-            }
-            Ok(MetricHandle::HistogramVec(histogram_vec))
-        }
-        Err(_) => Err(anyhow::anyhow!("Error creating histogram")),
-    }
-}
-
-/// Create a metric.
-///
-/// # Parameter
-/// - `system_ptr`: Pointer to the previously created TelemetrySystem.
-/// - `metric_type`: 0 for Counter, 1 for Gauge, 2 for Histogram.
-/// - `name`, `help`: The name and description of the metric.
-/// - `label_names_ptr` and `label_names_len`: If not null, a metric with labels
-///   will be created; only label names are required at creation time.
-///
-/// # Return
-/// Returns a pointer to MetricHandle on success, otherwise returns null.
-#[no_mangle]
-#[allow(clippy::not_unsafe_ptr_arg_deref)]
-pub extern "C" fn ten_metric_create(
-    system_ptr: *mut TelemetrySystem,
-    metric_type: u32, // 0=Counter, 1=Gauge, 2=Histogram
-    name: *const c_char,
-    help: *const c_char,
-    label_names_ptr: *const *const c_char,
-    label_names_len: usize,
-) -> *mut MetricHandle {
-    debug_assert!(!system_ptr.is_null(), "System pointer is null");
-    debug_assert!(!name.is_null(), "Name is null for metric creation");
-    debug_assert!(!help.is_null(), "Help is null for metric creation");
-
-    if system_ptr.is_null() || name.is_null() || help.is_null() {
-        return ptr::null_mut();
-    }
-
-    let system = unsafe { &mut *system_ptr };
-
-    let name_str = match unsafe { CStr::from_ptr(name) }.to_str() {
-        Ok(s) => s,
-        Err(_) => return ptr::null_mut(),
-    };
-    let help_str = match unsafe { CStr::from_ptr(help) }.to_str() {
-        Ok(s) => s,
-        Err(_) => return ptr::null_mut(),
-    };
-
-    let label_names_owned =
-        match convert_label_names(label_names_ptr, label_names_len) {
-            Some(v) => v,
-            None => return ptr::null_mut(),
-        };
-    let label_names: Vec<&str> =
-        label_names_owned.iter().map(|s| s.as_str()).collect();
-
-    let metric_handle = match metric_type {
-        0 => {
-            // Counter.
-            if label_names.is_empty() {
-                match create_metric_counter(system, name_str, help_str) {
-                    Ok(metric) => metric,
-                    Err(_) => return ptr::null_mut(),
-                }
-            } else {
-                match create_metric_counter_with_labels(
-                    system,
-                    name_str,
-                    help_str,
-                    &label_names,
-                ) {
-                    Ok(metric) => metric,
-                    Err(_) => return ptr::null_mut(),
-                }
+        match server_thread_handle.join() {
+            Ok(_) => eprintln!("Endpoint server thread joined successfully"),
+            Err(e) => {
+                eprintln!("Error joining endpoint server thread: {:?}", e)
             }
         }
-        1 => {
-            // Gauge.
-            if label_names.is_empty() {
-                match create_metric_gauge(system, name_str, help_str) {
-                    Ok(metric) => metric,
-                    Err(_) => return ptr::null_mut(),
-                }
-            } else {
-                match create_metric_gauge_with_labels(
-                    system,
-                    name_str,
-                    help_str,
-                    &label_names,
-                ) {
-                    Ok(metric) => metric,
-                    Err(_) => return ptr::null_mut(),
-                }
-            }
-        }
-        2 => {
-            // Histogram.
-            if label_names.is_empty() {
-                match create_metric_histogram(system, name_str, help_str) {
-                    Ok(metric) => metric,
-                    Err(_) => return ptr::null_mut(),
-                }
-            } else {
-                match create_metric_histogram_with_labels(
-                    system,
-                    name_str,
-                    help_str,
-                    &label_names,
-                ) {
-                    Ok(metric) => metric,
-                    Err(_) => return ptr::null_mut(),
-                }
-            }
-        }
-        _ => return ptr::null_mut(),
-    };
-
-    Box::into_raw(Box::new(metric_handle))
-}
-
-/// Release a metric handle created by `ten_metric_create`.
-#[no_mangle]
-#[allow(clippy::not_unsafe_ptr_arg_deref)]
-pub extern "C" fn ten_metric_destroy(metric_ptr: *mut MetricHandle) {
-    debug_assert!(!metric_ptr.is_null(), "Metric pointer is null");
-
-    if metric_ptr.is_null() {
-        return;
+    } else {
+        eprintln!("No thread handle available for the endpoint server");
     }
 
-    unsafe {
-        let _ = Box::from_raw(metric_ptr);
-    }
-}
-
-#[no_mangle]
-#[allow(clippy::not_unsafe_ptr_arg_deref)]
-pub extern "C" fn ten_metric_counter_inc(
-    metric_ptr: *mut MetricHandle,
-    label_values_ptr: *const *const c_char,
-    label_values_len: usize,
-) {
-    debug_assert!(!metric_ptr.is_null(), "Metric pointer is null");
-
-    if metric_ptr.is_null() {
-        return;
-    }
-
-    let metric = unsafe { &mut *metric_ptr };
-    match metric {
-        MetricHandle::Counter(ref counter) => {
-            counter.inc();
-        }
-        MetricHandle::CounterVec(ref counter_vec) => {
-            let values_owned = match convert_label_values(
-                label_values_ptr,
-                label_values_len,
-            ) {
-                Some(v) => v,
-                None => return,
-            };
-            let label_refs: Vec<&str> =
-                values_owned.iter().map(|s| s.as_str()).collect();
-            if let Ok(counter) =
-                counter_vec.get_metric_with_label_values(&label_refs)
-            {
-                counter.inc();
-            }
-        }
-        _ => {}
-    }
-}
-
-#[no_mangle]
-#[allow(clippy::not_unsafe_ptr_arg_deref)]
-pub extern "C" fn ten_metric_counter_add(
-    metric_ptr: *mut MetricHandle,
-    value: f64,
-    label_values_ptr: *const *const c_char,
-    label_values_len: usize,
-) {
-    debug_assert!(!metric_ptr.is_null(), "Metric pointer is null");
-
-    if metric_ptr.is_null() {
-        return;
-    }
-
-    let metric = unsafe { &mut *metric_ptr };
-    match metric {
-        MetricHandle::Counter(ref counter) => {
-            counter.inc_by(value);
-        }
-        MetricHandle::CounterVec(ref counter_vec) => {
-            let values_owned = match convert_label_values(
-                label_values_ptr,
-                label_values_len,
-            ) {
-                Some(v) => v,
-                None => return,
-            };
-            let label_refs: Vec<&str> =
-                values_owned.iter().map(|s| s.as_str()).collect();
-            if let Ok(counter) =
-                counter_vec.get_metric_with_label_values(&label_refs)
-            {
-                counter.inc_by(value);
-            }
-        }
-        _ => {}
-    }
-}
-
-#[no_mangle]
-#[allow(clippy::not_unsafe_ptr_arg_deref)]
-pub extern "C" fn ten_metric_gauge_set(
-    metric_ptr: *mut MetricHandle,
-    value: f64,
-    label_values_ptr: *const *const c_char,
-    label_values_len: usize,
-) {
-    debug_assert!(!metric_ptr.is_null(), "Metric pointer is null");
-
-    if metric_ptr.is_null() {
-        return;
-    }
-
-    let metric = unsafe { &mut *metric_ptr };
-    match metric {
-        MetricHandle::Gauge(ref gauge) => {
-            gauge.set(value);
-        }
-        MetricHandle::GaugeVec(ref gauge_vec) => {
-            let values_owned = match convert_label_values(
-                label_values_ptr,
-                label_values_len,
-            ) {
-                Some(v) => v,
-                None => return,
-            };
-            let label_refs: Vec<&str> =
-                values_owned.iter().map(|s| s.as_str()).collect();
-            if let Ok(gauge) =
-                gauge_vec.get_metric_with_label_values(&label_refs)
-            {
-                gauge.set(value);
-            }
-        }
-        _ => {}
-    }
-}
-
-#[no_mangle]
-#[allow(clippy::not_unsafe_ptr_arg_deref)]
-pub extern "C" fn ten_metric_gauge_inc(
-    metric_ptr: *mut MetricHandle,
-    label_values_ptr: *const *const c_char,
-    label_values_len: usize,
-) {
-    debug_assert!(!metric_ptr.is_null(), "Metric pointer is null");
-
-    if metric_ptr.is_null() {
-        return;
-    }
-
-    let metric = unsafe { &mut *metric_ptr };
-    match metric {
-        MetricHandle::Gauge(ref gauge) => {
-            gauge.inc();
-        }
-        MetricHandle::GaugeVec(ref gauge_vec) => {
-            let values_owned = match convert_label_values(
-                label_values_ptr,
-                label_values_len,
-            ) {
-                Some(v) => v,
-                None => return,
-            };
-            let label_refs: Vec<&str> =
-                values_owned.iter().map(|s| s.as_str()).collect();
-            if let Ok(gauge) =
-                gauge_vec.get_metric_with_label_values(&label_refs)
-            {
-                gauge.inc();
-            }
-        }
-        _ => {}
-    }
-}
-
-#[no_mangle]
-#[allow(clippy::not_unsafe_ptr_arg_deref)]
-pub extern "C" fn ten_metric_gauge_dec(
-    metric_ptr: *mut MetricHandle,
-    label_values_ptr: *const *const c_char,
-    label_values_len: usize,
-) {
-    debug_assert!(!metric_ptr.is_null(), "Metric pointer is null");
-
-    if metric_ptr.is_null() {
-        return;
-    }
-
-    let metric = unsafe { &mut *metric_ptr };
-    match metric {
-        MetricHandle::Gauge(ref gauge) => {
-            gauge.dec();
-        }
-        MetricHandle::GaugeVec(ref gauge_vec) => {
-            let values_owned = match convert_label_values(
-                label_values_ptr,
-                label_values_len,
-            ) {
-                Some(v) => v,
-                None => return,
-            };
-            let label_refs: Vec<&str> =
-                values_owned.iter().map(|s| s.as_str()).collect();
-            if let Ok(gauge) =
-                gauge_vec.get_metric_with_label_values(&label_refs)
-            {
-                gauge.dec();
-            }
-        }
-        _ => {}
-    }
-}
-
-#[no_mangle]
-#[allow(clippy::not_unsafe_ptr_arg_deref)]
-pub extern "C" fn ten_metric_histogram_observe(
-    metric_ptr: *mut MetricHandle,
-    value: f64,
-    label_values_ptr: *const *const c_char,
-    label_values_len: usize,
-) {
-    debug_assert!(!metric_ptr.is_null(), "Metric pointer is null");
-
-    if metric_ptr.is_null() {
-        return;
-    }
-
-    let metric = unsafe { &mut *metric_ptr };
-    match metric {
-        MetricHandle::Histogram(ref histogram) => {
-            histogram.observe(value);
-        }
-        MetricHandle::HistogramVec(ref histogram_vec) => {
-            let values_owned = match convert_label_values(
-                label_values_ptr,
-                label_values_len,
-            ) {
-                Some(v) => v,
-                None => return,
-            };
-            let label_refs: Vec<&str> =
-                values_owned.iter().map(|s| s.as_str()).collect();
-            if let Ok(histogram) =
-                histogram_vec.get_metric_with_label_values(&label_refs)
-            {
-                histogram.observe(value);
-            }
-        }
-        _ => {}
-    }
+    // The system will be automatically dropped here, cleaning up all resources.
 }
